Made various changes to bin/ folder
Added an --include option to `sumlines` which allows counting of only included
files based on a regex pattern.

Removed the `markdown-to-html` script

Added the new script `youtube-dl-batch` which takes files as arguments and which
are read line by line and contain `youtube-dl` options.

Completely rewrote the `trash` script in python.